feat: bot server added event
