+++ packages/guilded.js/lib/constants.ts
@@ -29,6 +29,7 @@ export const constants = {
         DOC_UPDATED: "docUpdated",
         DOC_DELETED: "docDeleted",
         SERVER_CREATED: "serverCreated",
+        BOT_SERVER_CREATED: "botServerCreated",
         MESSAGE_REACTION_CREATED: "messageReactionCreated",
         MESSAGE_REACTION_DELETED: "messageReactionDeleted",
         FORUM_TOPIC_CREATED: "forumTopicCreated",
